Adjust data type to size_t

+++ src/algorithms/sort/bubble sort/main.c
@@ -6,10 +6,10 @@
 
 int main()
 {
-    int array[] = {7, 1, 326, 832, 225, 32323, 45, 1, 36, 43534, 34534, 2};
+    size_t array[] = {7, 1, 326, 832, 225, 32323, 45, 1, 36, 43534, 34534, 2};
     bubble_sort(array, LENGTH);
         for (int i = 0; i < LENGTH; ++i)
     {
-        printf("%d\n", array[i]);
+        printf("%zd\n", array[i]);
     }
 }
